Implement Navigation and Data Binding
-Move Activity code to DriveLoginFragment
-Change MvvmActivity extension to MvvmFragment
-Add binding references in the fragment
-Change HerdrActivity to only set activity_herdr.xml view
-Connect navigation to activity_herdr.xml
